Canonical link: https://commits.webkit.org/188554.91@safari-604.1.21-branch
git-svn-id: https://svn.webkit.org/repository/webkit/branches/safari-604.1.21-branch@217662 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/webkitpy/port/simulator_process.py
@@ -120,6 +120,6 @@ class SimulatorProcess(ServerProcess):
         self._proc = SimulatorProcess.Popen(self._pid, stdin, stdout, stderr, self._target_host)
 
     def stop(self, timeout_secs=3.0):
-        if self._proc:
+        if self._proc and self._proc.pid:
             self._target_host.executive.kill_process(self._proc.pid)
         return super(SimulatorProcess, self).stop(timeout_secs)
